feat(ansi): use O(n) ansi aware coloring hard reset
On previous versions, the chromashift would use \x1b[0m to hard reset at
the end of the style. This creates issue when nesting if a bigger match
use 'underline' and a nested smaller caller uses other style hard reset
for that style will also hard reset the 'underline'. Now the ansi
colorizer is aware of context by a stack based implementation which
should solve this.